fix(the-storm): bind mechanics structures to their signs so no block can be duplicated
Why: review found two duplication bugs. A double slab placed as one
stored block came back as two, and opening a structure could pop a sign
off mid-toggle so its stock dropped and was then written again. Radius
pooling also let a sign drain or fill stock on other people's signs.

What:
- Structures are bound when their sign is written, in the sign's data.
  Bridges and doors link exactly two end signs to each other; the first
  end keeps the stock, the second holds none. A gate sign remembers its
  column tops and material. Every use checks the structure still matches
  (partner, material, end blocks, column tops) and refuses otherwise; no
  sign ever re-targets, and no other sign is read or written.
- Writing a structure sign that holds blocks is refused; rewriting a
  sign as plain text forgets its binding but keeps (and drops) its stock.
- A toggle never moves a block that holds anything up (signs, torches,
  levers, rails, carpets, item frames...), stores the new stock before
  any block moves, and needs the creator's BREAK and BUILD on every cell.
  Linking also needs BUILD at the other end's sign.
- Structure materials must be solid blocks that are always one item:
  slabs, candles, sea pickles, turtle eggs, snow, flower beds, doors and
  beds are refused at enable; slabs left the shipped block lists.
- Redstone needs the creator's INTERACT at the powering block. [Bounce]
  needs the creator's BUILD at the head, launches players and passive
  creatures only where Protection.checkHarm allows the creator (at the
  piston) to harm them, hostile mobs always, and other entities only from
  land the creator may build on.
- Structures move at most once per structureCooldownTicks (min 20).
- Glass and bookshelf self-drops use BlockDropItemEvent; searches never
  load chunks or cross the world border; stock counts are longs capped
  at one million.

Verification:
- mise exec -- gradle build (mechanics: 201 tests, none skipped)
- Domain tests over a grid that pops signs like physics: links, anchor
  and partner mismatches, a foreign gate sign left untouched, a
  supporting cell refusing to open, and blocks conserved across toggles
  with the stock sign broken mid-cycle. MockBukkit tests refuse slabs,
  candles, sea pickles, turtle eggs, snow and doors.
- Not run (needs the e2e server): physics pops, hanging-entity support
  detection, redstone, pistons and bounce protection, drop events.

+++ packages/the-storm/plugin/modules/mechanics/src/main/java/com/shepherdjerred/thestorm/mechanics/adapter/paper/Items.java
@@ -10,9 +10,9 @@ final class Items {
   private Items() {}
 
   /** Gives {@code count} of {@code kind}, dropping what does not fit at the player's feet. */
-  static void give(Player player, ItemStack kind, int count) {
+  static void give(Player player, ItemStack kind, long count) {
     for (var remaining = count; remaining > 0; ) {
-      var stack = kind.asQuantity(Math.min(remaining, kind.getMaxStackSize()));
+      var stack = kind.asQuantity((int) Math.min(remaining, kind.getMaxStackSize()));
       remaining -= stack.getAmount();
       player
           .getInventory()
@@ -23,9 +23,9 @@ final class Items {
   }
 
   /** Drops {@code count} of {@code kind} at {@code location}, in full stacks. */
-  static void drop(Location location, ItemStack kind, int count) {
+  static void drop(Location location, ItemStack kind, long count) {
     for (var remaining = count; remaining > 0; ) {
-      var stack = kind.asQuantity(Math.min(remaining, kind.getMaxStackSize()));
+      var stack = kind.asQuantity((int) Math.min(remaining, kind.getMaxStackSize()));
       remaining -= stack.getAmount();
       location.getWorld().dropItemNaturally(location, stack);
     }
